+++ deps/vecmath/src/Matrix2f.cpp
@@ -7,12 +7,12 @@
 
 #include "Vector2f.h"
 
-Matrix2f::Matrix2f(float fill) {
+Matrix2f::Matrix2f(double fill) {
     for (int i = 0; i < 4; ++i)
         m_elements[i] = fill;
 }
 
-Matrix2f::Matrix2f(float m00, float m01, float m10, float m11) {
+Matrix2f::Matrix2f(double m00, double m01, double m10, double m11) {
     m_elements[0] = m00;
     m_elements[1] = m10;
 
@@ -31,20 +31,20 @@ Matrix2f::Matrix2f(const Vector2f& v0, const Vector2f& v1, bool setColumns) {
 }
 
 Matrix2f::Matrix2f(const Matrix2f& rm) {
-    memcpy(m_elements, rm.m_elements, 2 * sizeof(float));
+    memcpy(m_elements, rm.m_elements, 2 * sizeof(double));
 }
 
 Matrix2f& Matrix2f::operator=(const Matrix2f& rm) {
     if (this != &rm)
-        memcpy(m_elements, rm.m_elements, 2 * sizeof(float));
+        memcpy(m_elements, rm.m_elements, 2 * sizeof(double));
     return *this;
 }
 
-const float& Matrix2f::operator()(int i, int j) const {
+const double& Matrix2f::operator()(int i, int j) const {
     return m_elements[j * 2 + i];
 }
 
-float& Matrix2f::operator()(int i, int j) {
+double& Matrix2f::operator()(int i, int j) {
     return m_elements[j * 2 + i];
 }
 
@@ -76,15 +76,15 @@ void Matrix2f::setCol(int j, const Vector2f& v) {
     m_elements[colStart + 1] = v.y();
 }
 
-float Matrix2f::determinant() {
+double Matrix2f::determinant() {
     return Matrix2f::determinant2x2 (
         m_elements[0], m_elements[2],
         m_elements[1], m_elements[3]
     );
 }
 
-Matrix2f Matrix2f::inverse(bool* pbIsSingular, float epsilon) {
-    float determinant = m_elements[0] * m_elements[3] - m_elements[2] * m_elements[1];
+Matrix2f Matrix2f::inverse(bool* pbIsSingular, double epsilon) {
+    double determinant = m_elements[0] * m_elements[3] - m_elements[2] * m_elements[1];
 
     bool isSingular = (fabs(determinant) < epsilon);
     if (isSingular) {
@@ -95,7 +95,7 @@ Matrix2f Matrix2f::inverse(bool* pbIsSingular, float epsilon) {
         if (pbIsSingular != NULL)
             *pbIsSingular = false;
 
-        float reciprocalDeterminant = 1.0f / determinant;
+        double reciprocalDeterminant = 1.0 / determinant;
         return Matrix2f(
             m_elements[3] * reciprocalDeterminant, -m_elements[2] * reciprocalDeterminant,
             -m_elements[1] * reciprocalDeterminant, m_elements[0] * reciprocalDeterminant
@@ -104,8 +104,8 @@ Matrix2f Matrix2f::inverse(bool* pbIsSingular, float epsilon) {
 }
 
 void Matrix2f::transpose() {
-    float m01 = (*this)(0, 1);
-    float m10 = (*this)(1, 0);
+    double m01 = (*this)(0, 1);
+    double m10 = (*this)(1, 0);
 
     (*this)(0, 1) = m10;
     (*this)(1, 0) = m01;
@@ -118,7 +118,7 @@ Matrix2f Matrix2f::transposed() const {
     );
 }
 
-Matrix2f::operator float*() {
+Matrix2f::operator double*() {
     return m_elements;
 }
 
@@ -129,7 +129,7 @@ void Matrix2f::print() {
 }
 
 // Static method
-float Matrix2f::determinant2x2(float m00, float m01, float m10, float m11) {
+double Matrix2f::determinant2x2(double m00, double m01, double m10, double m11) {
     return (m00 * m11 - m01 * m10);
 }
 
@@ -152,9 +152,9 @@ Matrix2f Matrix2f::identity() {
 }
 
 // Static method
-Matrix2f Matrix2f::rotation(float degrees) {
-    float c = cos(degrees);
-    float s = sin(degrees);
+Matrix2f Matrix2f::rotation(double degrees) {
+    double c = cos(degrees);
+    double s = sin(degrees);
 
     return Matrix2f(
         c, -s,
@@ -166,7 +166,7 @@ Matrix2f Matrix2f::rotation(float degrees) {
 // Operators
 //////////////////////////////////////////////////////////////////////////
 
-Matrix2f operator*(float f, const Matrix2f& m) {
+Matrix2f operator*(double f, const Matrix2f& m) {
     Matrix2f output;
 
     for (int i = 0; i < 2; ++i)
@@ -176,7 +176,7 @@ Matrix2f operator*(float f, const Matrix2f& m) {
     return output;
 }
 
-Matrix2f operator*(const Matrix2f& m, float f) {
+Matrix2f operator*(const Matrix2f& m, double f) {
     return f * m;
 }
